requirements.txt
pandas
numpy
datamol
rdkit-pypi
mordred
molfeat
morfeus-ml
dscribe
splito
ray
scikit-learn
xgboost
lightgbm
catboost
ngboost
mlxtend
yellowbrick
gpflow
gpy
matplotlib
seaborn
mpl_scatter_density
shap
